Disabled-Door-Button/requirements.txt
numpy==1.26.4
pyrtlsdr==0.2.93
requests==2.34.2